crime pattern analysis
The project is a crime arrest pattern analysis of Indian cities. It uses a given dataset to identify the top cities with the most arrests and analyzes the distribution of arrests by age and gender. The project's goal is to convert raw data into clear, visual insights to understand key trends.